requirements.txt
# Flask
transformers
torch==2.8.0
sentencepiece
flask_sqlalchemy==3.1.1
flask_login==0.6.3
werkzeug==3.1.3
wtforms==3.2.0
flask_wtf==1.2.2
flask_migrate==4.1.0
